requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.5
djangorestframework==3.12.4
idna==3.2
pycoingecko==2.2.0
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
